version 2.1.8

+++ lib/Classes/FilesManager.hpp
@@ -11,8 +11,8 @@ class FilesManager
 public:
 
   FilesManager(){};
-  FilesManager(std::string const & folder, int nb_files = -1){addFolder(folder,nb_files);};
-  FilesManager(Path & folder, int nb_files = -1){addFolder(folder,nb_files);};
+  FilesManager(std::string const & folder, long nb_files = -1){addFolder(folder,nb_files);};
+  FilesManager(Path & folder, long nb_files = -1){addFolder(folder,nb_files);};
 
 
   bool nextFileName(std::string & filename, size_t const & step = 1);
@@ -20,7 +20,7 @@ public:
   // Adds either a single file or reads a .list containing a list of files
   virtual bool addFiles     (std::string const & _filename        );
   // Adds a given number of files with a .root or .fast inside the given folder (by default all the files, or the nb_files first ones)
-  virtual bool addFolder    (std::string folder, int nb_files = -1);
+  virtual bool addFolder    (std::string folder, long nb_files = -1);
   virtual void flushFiles   ();
   void   Print        () { for (auto const & file : m_listFiles) print(file);}
   void   printFolders () { for (auto const & folder : m_listFolder) print(folder);}
@@ -113,7 +113,7 @@ bool FilesManager::addFiles(std::string const & _filename)
   << "unkown..." << std::endl << "Abort..." << std::endl;return false;}
 }
 
-bool FilesManager::addFolder(std::string _foldername, int _nb_files)
+bool FilesManager::addFolder(std::string _foldername, long _nb_files)
 {
   if (_foldername.back() != '/')
   {
@@ -143,7 +143,7 @@ bool FilesManager::addFolder(std::string _foldername, int _nb_files)
   if (listfile.size() > 0)
   {
     std::sort(listfile.begin(), listfile.end());// Sorts the entries
-    if (_nb_files> static_cast<int>(listfile.size()) || _nb_files == -1) _nb_files = listfile.size();//Sets the correct number of files to keep
+    if (_nb_files> static_cast<long>(listfile.size()) || _nb_files == -1) _nb_files = listfile.size();//Sets the correct number of files to keep
     ListFiles cut_listfile (listfile.begin(), listfile.begin()+_nb_files);// Take the nb_files first files of the folder
     for (auto const & file : cut_listfile) m_listFilesInFolder[_foldername].emplace_back(file);
     if (m_listFiles.size() == 0) m_listFiles = cut_listfile;// Set cut_listfile to be the global list of files
